+++ Tests/KeyboardKitTests/Layout/Providers/StandardKeyboardLayoutProviderTests.swift
@@ -16,6 +16,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
         
         var inputSetProvider: MockInputSetProvider!
         var context: KeyboardContext!
+        var provider: StandardKeyboardLayoutProvider!
         
         beforeEach {
             context = KeyboardContext()
@@ -23,11 +24,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             inputSetProvider.alphabeticInputSetValue = AlphabeticInputSet(rows: [["a", "b", "c"], ["a", "b", "c"], ["a", "b", "c"]].map(InputSetRow.init))
             inputSetProvider.numericInputSetValue = NumericInputSet(rows: [["1", "2", "3"], ["1", "2", "3"], ["1", "2", "3"]].map(InputSetRow.init))
             inputSetProvider.symbolicInputSetValue = SymbolicInputSet(rows: [[",", ".", "-"], [",", ".", "-"], [",", ".", "-"]].map(InputSetRow.init))
-        }
-
-        func provider(for device: DeviceType) -> StandardKeyboardLayoutProvider {
-            StandardKeyboardLayoutProvider(
-                device: device,
+            provider = StandardKeyboardLayoutProvider(
                 inputSetProvider: inputSetProvider,
                 dictationReplacement: .primary(.go))
         }
@@ -35,14 +32,14 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
         describe("keyboard layout provider for context") {
             
             it("is phone provider if context device is phone") {
-                let provider = provider(for: .phone)
-                let result = provider.layoutProvider
+                context.deviceType = .phone
+                let result = provider.keyboardLayoutProvider(for: context)
                 expect(result).to(be(provider.iPhoneProvider))
             }
             
             it("is pad provider if context device is pad") {
-                let provider = provider(for: .pad)
-                let result = provider.layoutProvider
+                context.deviceType = .pad
+                let result = provider.keyboardLayoutProvider(for: context)
                 expect(result).to(be(provider.iPadProvider))
             }
         }
@@ -50,7 +47,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
         describe("keyboard layout for context (just testing this one)") {
             
             it("is phone layout if context device is phone") {
-                let provider = provider(for: .phone)
+                context.deviceType = .phone
                 let layout = provider.keyboardLayout(for: context)
                 let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
                 let padLayout = provider.iPadProvider.keyboardLayout(for: context)
@@ -59,7 +56,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             }
             
             it("is pad layout if context device is pad") {
-                let provider = provider(for: .pad)
+                context.deviceType = .pad
                 let layout = provider.keyboardLayout(for: context)
                 let phoneLayout = provider.iPhoneProvider.keyboardLayout(for: context)
                 let padLayout = provider.iPadProvider.keyboardLayout(for: context)
@@ -72,7 +69,7 @@ class StandardKeyboardLayoutProviderTests: QuickSpec {
             
             it("changes the provider instance for all providers") {
                 let newProvider = MockInputSetProvider()
-                let provider = provider(for: .phone)
+                context.deviceType = .phone
                 provider.register(inputSetProvider: newProvider)
                 expect(provider.inputSetProvider).toNot(be(inputSetProvider))
                 expect(provider.inputSetProvider).to(be(newProvider))